+++ app/src/main/java/com/yaabelozerov/superfinancer/domain/model/Transaction.kt
@@ -2,7 +2,7 @@ package com.yaabelozerov.superfinancer.domain.model
 
 data class Transaction(
     val id: Long,
-    val valueInRubles: Double,
+    val valueInRubles: Long,
     val comment: String,
     val goalName: String,
     val timestamp: String
